[Refactor] 파일분리
