+++ SOPT36_Assignment/Home/Model/BaseballModel.swift
