+++ SOPT36_Assignment/Home/Model/LogoModel.swift
