+++ SOPT36_Assignment/Home/Model/PosterModel.swift
