--- SOPT36_Assignment/ViewModel/MainViewModel.swift
@@ -1,8 +1,0 @@
-//
-//  MainViewModel.swift
-//  SOPT36_Assignment
-//
-//  Created by OneTen on 6/3/25.
-//
-
-import Foundation
